requirements.txt
annotated-types==0.7.0
anyio==4.8.0
atproto==0.0.56
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==43.0.3
dnspython==2.7.0
exceptiongroup==1.2.2
facebook-sdk==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
libipld==3.0.0
# Editable install with no version control (pubnwr==2.0.0)
-e /home/dmitri/NowWaveRadio/publish/Myriad_Claude/pubnwr
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pylast==5.3.0
pylistenbrainz==0.5.1
pytz==2024.2
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
watchdog==6.0.0
websockets==13.1
